Set m3 target in Makefile as phony
As we have the binary which is named as m3 in the main directory, the
target named m3 is not working properly and returns the message saying
'make: m3 is up to date'. Therefore, we would need to declare the m3
target in the Makefile as phony to avoid this error by force rebuilding
even when a file named m3 is present in the directory.