Adding support for custom ca certificates
If required, we can load custom ca certificates into the folder with the same name,
certificates will be installed when building the mkube image.
This solves the problem of using a kms with a certificate issued by a custom ca.
Note that operator still need to create a configmap in k8s based on the ca certificate and
define the KMS_CA_CERT_CONFIG_MAP and KMS_CA_CERT_FILE_NAME variables so kes
can use the custom ca certificate